+++ test/main/Main.hs
@@ -13,6 +13,7 @@ import Test.Calendar.Week
 import Test.Clock.Conversion
 import Test.Clock.Resolution
 import Test.Clock.TAI
+import Test.Format.Compile ()
 import Test.Format.Format
 import Test.Format.ISO8601
 import Test.Format.ParseTime